requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv
langchain==0.1.16
langchain-community==0.0.37
langchain-core>=0.1.51,<0.2.0
langchain-text-splitters==0.0.1
sentence-transformers==2.6.1
huggingface-hub==0.19.4
faiss-cpu==1.7.4
transformers==4.36.2
pdfplumber==0.10.3
torch==2.1.0
numpy==1.26.4

groq